+++ hapro/rootfs/usr/src/webApiControllers/backupController.ts
@@ -4,12 +4,12 @@ async function getBackups() {
   try {
     const response = await helpers.doSupervisorRequest("/backups");
     return new Response(
-      JSON.stringify({ StatusCode: 200, data: response.data.backups })
+      JSON.stringify({ StatusCode: 200, data: response.data.backups }),
     );
   } catch (error) {
     console.error(error);
     return new Response(
-      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" })
+      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
     );
   }
 }
@@ -17,7 +17,7 @@ async function getBackups() {
 async function getBackupInfo(backupId) {
   try {
     const response = await helpers.doSupervisorRequest(
-      `/backups/${backupId}/info`
+      `/backups/${backupId}/info`,
     );
     const returnObject = {
       slug: response.data.slug,
@@ -34,12 +34,12 @@ async function getBackupInfo(backupId) {
       },
     };
     return new Response(
-      JSON.stringify({ StatusCode: 200, data: returnObject })
+      JSON.stringify({ StatusCode: 200, data: returnObject }),
     );
   } catch (error) {
     console.error(error);
     return new Response(
-      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" })
+      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
     );
   }
 }
@@ -53,13 +53,13 @@ async function downloadBackup(backupId) {
         headers: {
           Authorization: `Bearer ${Bun.env.SUPERVISOR_TOKEN}`,
         },
-      }
+      },
     );
     return response;
   } catch (error) {
     console.error(error);
     return new Response(
-      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" })
+      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
     );
   }
 }
@@ -74,7 +74,7 @@ async function uploadBackup(req) {
     const response = await fetch("http://supervisor/backups/new/upload", {
       method: "POST",
       headers: {
-        Authorization: `Bearer ${Bun.env.SUPERVISOR_TOKEN}`
+        Authorization: `Bearer ${Bun.env.SUPERVISOR_TOKEN}`,
       },
       body: formData,
     });
@@ -84,14 +84,20 @@ async function uploadBackup(req) {
     }
 
     console.debug("Backup uploaded successfully");
-    return new Response(JSON.stringify({ StatusCode: 200, Message: "Backup uploaded successfully" }), {
-      status: 200,
-    });
+    return new Response(
+      JSON.stringify({
+        StatusCode: 200,
+        Message: "Backup uploaded successfully",
+      }),
+      {
+        status: 200,
+      },
+    );
   } catch (error) {
     console.error("Error uploading backup:", error);
     return new Response(
       JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
-      { status: 500 }
+      { status: 500 },
     );
   }
 }
@@ -100,13 +106,23 @@ async function deleteBackup(backupId) {
   try {
     const response = await helpers.doSupervisorRequest(
       `/backups/${backupId}`,
-      "DELETE"
+      "DELETE",
     );
-    return new Response(JSON.stringify({ StatusCode: response.result == "ok" ? 200 : response.message == "Backup does not exist" ? 404 : 500, Message: response.message }));
+    return new Response(
+      JSON.stringify({
+        StatusCode:
+          response.result == "ok"
+            ? 200
+            : response.message == "Backup does not exist"
+              ? 404
+              : 500,
+        Message: response.message,
+      }),
+    );
   } catch (error) {
     console.error(error);
     return new Response(
-      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" })
+      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
     );
   }
 }
@@ -114,53 +130,192 @@ async function deleteBackup(backupId) {
 async function restoreBackup(backupId, backupPassword = null) {
   try {
     const backups = await helpers.doSupervisorRequest("/backups");
-    const backup = backups.data.backups.find(backup => backup.slug === backupId);
-    if(backup === undefined)
-      return new Response(JSON.stringify({ StatusCode: 404, Message: "Backup not found" }));
-    if(backup.protected && backupPassword === null)
-      return new Response(JSON.stringify({ StatusCode: 401, Message: "Backup is password protected" }));
-    if(backup.type === "full") {
+    const backup = backups.data.backups.find(
+      (backup) => backup.slug === backupId,
+    );
+    if (backup === undefined)
+      return new Response(
+        JSON.stringify({ StatusCode: 404, Message: "Backup not found" }),
+      );
+    if (backup.protected && backupPassword === null)
+      return new Response(
+        JSON.stringify({
+          StatusCode: 401,
+          Message: "Backup is password protected",
+        }),
+      );
+    if (backup.type === "full") {
       const response = await helpers.doSupervisorRequest(
         `/backups/${backupId}/restore/full`,
         "POST",
-        { background: true }
+        { background: true },
       );
-      return new Response(JSON.stringify({ StatusCode: 200, Message: "Restore started", Data: response.data.job_id }));
+      return new Response(
+        JSON.stringify({
+          StatusCode: 200,
+          Message: "Restore started",
+          Data: response.data.job_id,
+        }),
+      );
     } else {
       const response = await helpers.doSupervisorRequest(
         `/backups/${backupId}/restore/partial`,
         "POST",
-        { background: true, homeassistant: backup.content.homeassistant, addons: backup.content.addons, folders: backup.content.folders }
+        {
+          background: true,
+          homeassistant: backup.content.homeassistant,
+          addons: backup.content.addons,
+          folders: backup.content.folders,
+        },
       );
-      return new Response(JSON.stringify({ StatusCode: 200, Message: "Restore started", Data: response.data.job_id }));
+      return new Response(
+        JSON.stringify({
+          StatusCode: 200,
+          Message: "Restore started",
+          Data: response.data.job_id,
+        }),
+      );
     }
   } catch (error) {
     console.error(error);
     return new Response(
-      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" })
+      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
     );
   }
 }
 
 async function backupStatus(jobId) {
   try {
-    const response = await helpers.doSupervisorRequest(
-      `/jobs/info`
+    const response = await helpers.doSupervisorRequest(`/jobs/info`);
+    const job = response.data.jobs.find((job) => job.uuid === jobId);
+    if (job === undefined)
+      return new Response(
+        JSON.stringify({ StatusCode: 404, Message: "Job not found" }),
+      );
+    if (job.done && job.errors.length > 0)
+      return new Response(
+        JSON.stringify({
+          StatusCode: 500,
+          Message: "Backup failed",
+          Data: job.errors,
+        }),
+      );
+    if (job.done)
+      return new Response(
+        JSON.stringify({ StatusCode: 200, Message: "Backup completed" }),
+      );
+    return new Response(
+      JSON.stringify({ StatusCode: 100, Message: "Backup in progress" }),
     );
-    const job = response.data.jobs.find(job => job.uuid === jobId);
-    if(job === undefined)
-      return new Response(JSON.stringify({ StatusCode: 404, Message: "Job not found" }));
-    if(job.done && job.errors.length > 0)
-      return new Response(JSON.stringify({ StatusCode: 500, Message: "Backup failed", Data: job.errors }));
-    if(job.done)
-      return new Response(JSON.stringify({ StatusCode: 200, Message: "Backup completed" }));
-    return new Response(JSON.stringify({ StatusCode: 100, Message: "Backup in progress" }));
   } catch (error) {
     console.error(error);
     return new Response(
-      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" })
+      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
     );
   }
 }
 
-export { getBackups, getBackupInfo, downloadBackup, uploadBackup, deleteBackup, restoreBackup, backupStatus };
+async function getBackupEmergencyKey() {
+  // read /homeassistant/.storage/backup and parse the json to get the key
+  try {
+    const backupFile = Bun.file("/homeassistant/.storage/backup");
+    const backupFileText = await backupFile.text();
+    const backupFileContent = JSON.parse(backupFileText);
+    const key = backupFileContent.data.config.create_backup.password;
+    if (key === undefined)
+      return new Response(
+        JSON.stringify({
+          StatusCode: 404,
+          Message: "Backup emergency key not found",
+        }),
+      );
+
+    return new Response(
+      JSON.stringify({
+        StatusCode: 200,
+        Message: "Backup emergency key retrieved successfully",
+        Data: key,
+      }),
+    );
+  } catch (error) {
+    console.error(
+      "Error reading backup emergency key:",
+      error instanceof Error ? error.message : error,
+    );
+    return new Response(
+      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
+    );
+  }
+}
+
+async function getBackupConfig() {
+  try {
+    const backupFile = Bun.file("/homeassistant/.storage/backup");
+    const backupHassioFile = Bun.file("/homeassistant/.storage/hassio");
+    const [backupFileText, backupHassioFileText] = await Promise.all([
+      backupFile.text(),
+      backupHassioFile.text(),
+    ]);
+    const backupFileContent = JSON.parse(backupFileText);
+    const backupFileData = backupFileContent.data.config;
+    const backupHassioFileContent = JSON.parse(backupHassioFileText);
+    const backupHassioFileData = backupHassioFileContent.data.update_config;
+    const backupConfig = {
+      automaticBackupsEnabled: backupFileData.automatic_backups_configured, //ha uses to determain a popup in the UI to ask the user to configure automatic backups, true once configured
+      content: backupFileData.automatic_backups_configured
+        ? {
+            includeAllAddons: backupFileData.create_backup.include_all_addons,
+            Addons: backupFileData.create_backup.include_addons, // filled if all_addons is false, if false and this is empty, no addons will be included
+            includeDatabase: backupFileData.create_backup.include_database, //history
+            Folders: backupFileData.create_backup.include_folders, // if empty, no folders will be included, media and share can be included here
+          }
+        : null,
+      retention: backupFileData.automatic_backups_configured
+        ? {
+            // both null = forever
+            copies: backupFileData.retention.copies, // int indicating how many backups to keep
+            days: backupFileData.retention.days, // int indicating how many days to keep backups
+          }
+        : null,
+      schedule: backupFileData.automatic_backups_configured
+        ? {
+            recurrence: backupFileData.schedule.recurrence, // never, daily, custom_days
+            days: backupFileData.schedule.days, // filled if recurrence is custom_days, values are the days of the week, e.g. ["mon", "tue"]
+            time: backupFileData.schedule.time, // time of the day null=System Optiomal, otherwise "HH:mm:00"
+          }
+        : null,
+      backupOnUpdate: {
+        homeassistant: backupHassioFileData.core_backup_before_update,
+        addons: backupHassioFileData.add_on_backup_before_update,
+        addonsRetain: backupHassioFileData.add_on_backup_retain_copies, // int of copies to keep
+      },
+    };
+    return new Response(
+      JSON.stringify({
+        StatusCode: 200,
+        Message: "Backup config retrieved successfully",
+        Data: backupConfig,
+      }),
+    );
+  } catch (error) {
+    console.error(
+      "Error reading backup config:",
+      error instanceof Error ? error.message : error,
+    );
+    return new Response(
+      JSON.stringify({ StatusCode: 500, Message: "Internal Server Error" }),
+    );
+  }
+}
+
+export {
+  getBackups,
+  getBackupInfo,
+  downloadBackup,
+  uploadBackup,
+  deleteBackup,
+  restoreBackup,
+  backupStatus,
+  getBackupEmergencyKey,
+  getBackupConfig,
+};